+++ lib/src/db/isoexpress/da_congestion_compact.dart
@@ -5,6 +5,10 @@ import 'dart:async';
 import 'package:collection/collection.dart';
 import 'package:date/date.dart';
 import 'package:elec_server/src/db/config.dart';
+import 'package:elec_server/src/db/isoexpress/da_lmp_hourly.dart';
+import 'package:elec_server/src/db/lib_prod_dbs.dart';
+import 'package:elec_server/src/db/webservices/da_lmp_hourly.dart';
+import 'package:more/comparator.dart';
 import 'package:more/ordering.dart';
 import 'package:timezone/timezone.dart';
 import 'package:tuple/tuple.dart';
@@ -21,22 +25,28 @@ class DaCongestionCompactArchive extends DailyIsoExpressReport {
         dbName: 'isoexpress',
         collectionName: 'da_congestion_compact');
     this.dbConfig = dbConfig;
-    dir ??= baseDir + 'PricingReports/DaLmpHourly/Raw/';
+    dir ??= '${baseDir}PricingReports/DaLmpHourly/Raw/';
     this.dir = dir;
     reportName = 'Day-Ahead Congestion Compact Archive';
   }
   final keys = {0, 0.01, 0.02};
 
-  @override
-  String getUrl(Date? asOfDate) =>
-      'https://www.iso-ne.com/static-transform/csv/histRpts/da-lmp/'
-          'WW_DALMP_ISO_' +
-      yyyymmdd(asOfDate) +
-      '.csv';
+  // @override
+  // String getUrl(Date? asOfDate) =>
+  //     'https://www.iso-ne.com/static-transform/csv/histRpts/da-lmp/WW_DALMP_ISO_${yyyymmdd(asOfDate)}.csv';
 
   @override
-  File getFilename(Date? asOfDate) =>
-      File(dir + 'WW_DALMP_ISO_' + yyyymmdd(asOfDate) + '.csv');
+  String getUrl(Date asOfDate) =>
+      'https://webservices.iso-ne.com/api/v1.1/hourlylmp/da/final/day/${yyyymmdd(asOfDate)}';
+
+  @override
+  File getFilename(Date asOfDate) {
+    if (asOfDate.isBefore(Date.utc(2022, 12, 22))) {
+      return File('${dir}WW_DALMP_ISO_${yyyymmdd(asOfDate)}.csv');
+    } else {
+      return File('${dir}WW_DALMP_ISO_${yyyymmdd(asOfDate)}.json');
+    }
+  }
 
   /// Return one document for one day for all ptids in the pool.
   /// Use rle to encode most common values.
@@ -55,69 +65,91 @@ class DaCongestionCompactArchive extends DailyIsoExpressReport {
   /// ```
   @override
   List<Map<String, dynamic>> processFile(File file) {
-    var _data = mis.readReportTabAsMap(file, tab: 0);
-    if (_data.isEmpty) return <Map<String, dynamic>>[];
-    var date = parseMmddyyy(_data.first['Date']);
 
-    // On some days, ISO duplicates the data for one (ptid, HE).  Make sure
-    // you only keep one!  See 2019-01-02 for example.
-    var aux = groupBy(
-            _data,
-            (Map row) =>
-                Tuple2(int.parse(row['Location ID']), row['Hour Ending']))
-        .map((key, value) => MapEntry(key, value.first));
-    var dataByPtids =
-        groupBy(aux.values, (Map row) => int.parse(row['Location ID']));
-    var _ptids = dataByPtids.keys.toList();
+    /// use the file processor from the DaLmpHourlyArchive
+    var damArchive = DaLmpHourlyArchive(dbConfig: dbConfig, dir: dir);
+    var docs = damArchive.processFile(file);
+    if (docs.isEmpty) return <Map<String, dynamic>>[];
 
-    // Initially, store the congestion in a List<List<num>> [ptid][hour].
-    // It gets transposed after the sorting.
-    var _congestion = <List<num>>[];
-    for (var ptid in dataByPtids.keys) {
-      _congestion.add(dataByPtids[ptid]!
-          .map((e) => e['Congestion Component'] as num)
-          .toList());
-    }
+    var ptids = docs.map((e) => e['ptid'] as int).toList();
 
-    /// insert the ptid index at position 0, so you can keep track of the
+    /// Initially, store the congestion in a List<List<num>> [ptid][hour].
+    /// It gets transposed after the sorting.
+    /// Insert the ptid index at position 0, so you can keep track of the
     /// ptid after you do the sorting.
-    for (var i = 0; i < _ptids.length; i++) {
-      _congestion[i].insert(0, i);
+    var congestion = <List<num>>[];
+    var i = 0;
+    for (var e in docs) {
+      congestion.add([i, ...e['congestion']]);
+      i++;
     }
+
+
+    // var _data = mis.readReportTabAsMap(file, tab: 0);
+    // if (_data.isEmpty) return <Map<String, dynamic>>[];
+    // var date = parseMmddyyy(_data.first['Date']);
+    //
+    // // On some days, ISO duplicates the data for one (ptid, HE).  Make sure
+    // // you only keep one!  See 2019-01-02 for example.
+    // var aux = groupBy(
+    //         _data,
+    //         (Map row) =>
+    //             Tuple2(int.parse(row['Location ID']), row['Hour Ending']))
+    //     .map((key, value) => MapEntry(key, value.first));
+    // var dataByPtids =
+    //     groupBy(aux.values, (Map row) => int.parse(row['Location ID']));
+    // var _ptids = dataByPtids.keys.toList();
+    //
+    // // Initially, store the congestion in a List<List<num>> [ptid][hour].
+    // // It gets transposed after the sorting.
+    // var _congestion = <List<num>>[];
+    // for (var ptid in dataByPtids.keys) {
+    //   _congestion.add(dataByPtids[ptid]!
+    //       .map((e) => e['Congestion Component'] as num)
+    //       .toList());
+    // }
+
+    // /// insert the ptid index at position 0, so you can keep track of the
+    // /// ptid after you do the sorting.
+    // for (var i = 0; i < _ptids.length; i++) {
+    //   _congestion[i].insert(0, i);
+    // }
 
     /// order the congestion data
-    var ordering = Ordering.natural<num>()
-        .onResultOf((List xs) => xs[1]) // sort by hour beginning 0
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[2]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[3]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[4]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[6]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[9]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[12]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[15]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[18]))
-        .compound(Ordering.natural<num>().onResultOf((List xs) => xs[21]));
-    ordering.sort(_congestion);
+    var ordering = naturalComparator<num>()
+        .onResultOf((List xs) => xs[1])
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[2]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[3]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[4]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[6]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[9]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[12]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[12]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[15]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[18]))
+      .thenCompare(naturalComparator<num>().onResultOf((List xs) => xs[21]));
+    ordering.sort(congestion);
 
     /// Transpose the _congestion matrix into a
     /// data matrix with index [hour][ptid]
-    var hoursCount = _congestion.first.length - 1;
+    var hoursCount = congestion.first.length - 1;
     var data = List.generate(
-        hoursCount, (i) => List<num>.generate(_ptids.length, (i) => 999.9));
+        hoursCount, (i) => List<num>.generate(ptids.length, (i) => 999.9));
     for (var i = 0; i < hoursCount; i++) {
-      for (var j = 0; j < _ptids.length; j++) {
-        data[i][j] = _congestion[j][i + 1];
+      for (var j = 0; j < ptids.length; j++) {
+        data[i][j] = congestion[j][i + 1];
       }
     }
 
     var out = {
-      'date': date.toString(),
-      'ptids': _congestion.map((e) => _ptids[e[0] as int]).toList(),
+      'date': docs.first['date'],
+      'ptids': congestion.map((e) => ptids[e[0] as int]).toList(),
       'congestion': data.map((List<num> e) => runLenghtEncode(e)).toList(),
     };
 
     return [out];
   }
+
 
   /// Recreate the collection from scratch.
   @override