+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## TODO:
 - 
 
+## Release 2023-01-02
+- Switch da_lmp_hourly in ISONE to the json webservices.  The public csv service is just not 
+reliable enough.  Stop inserting the 'hourBeginning' field from the documents in the collection.  
+It is not needed as all hours in the day are published so the info is redundant.  Save on storage 
+too.  Modify da_congestion_compact to work with json files too
+
 ## Release 2022-12-14
 - More work on retail offers.  Added MA residential offers.  Hope that the DB format has stabilized.